ability to see why CheckUserPassword failed
I'm trying to be able to detect why CheckUserPassword failed. This is so
we can support AD's ability to force a user to change their password on
the next login.
The AD server sends a "data code" in the message that isn't parsed by
OpenLDAP.
OpenLDAP does store the string and make's it available via
LDAP_OPT_DIAGNOSTIC_MESSAGE until the next error.
Example AD message for "user must change password":
`8009030C: LdapErr: DSID-0C090516, comment: AcceptSecurityContext error,
data 773, v3839`

data 773 tells us the code we need.

This pull requests adds the diagnostic message to the end of the error
string libadclient creates so I can parse the code out later in Go.

This pull request also adds a throw in adclient::checkUserPassword so it can be caught in
adclient_wrapper.go otherwise the error CheckUserPassword returns is
always nil.